Paging için IProductRepository ve ProductRepository tamamlandı. View de dahil olmak üzere paging ile alakalı her şey tamamlandı.

+++ Entity Layer/Product.cs
@@ -11,8 +11,8 @@ namespace Entity_Layer
         public int ProductId { get; set; }
         public string ProductName { get; set; }
         public string ProductDescription { get; set; }
-        public double UnitPrice { get; set; }
-        public int StockQuantity { get; set; }
+        public decimal UnitPrice { get; set; }
+        public short StockQuantity { get; set; }
         public int CategoryId { get; set; } 
         public Category Category { get; set; }
         public string CountryOfOrigin { get; set; }
